Fix vehicle giving in freeroam menu

+++ TDS_Server_DB/Entity/FreeroamDefaultVehicle.cs
@@ -6,7 +6,7 @@ namespace TDS_Server_DB.Entity
     public partial class FreeroamDefaultVehicle
     {
         public short VehicleTypeId { get; set; }
-        public int VehicleHash { get; set; }
+        public long VehicleHash { get; set; }
         public string Note { get; set; }
 
         public virtual FreeroamVehicleType VehicleType { get; set; }
